requirements.txt
Flask==3.0.0
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
gunicorn==21.2.0
Flask-Limiter==3.5.0
Pillow==11.3.0
